[ADD] votes passed

+++ internal/pkg/models/model_post.go
@@ -8,10 +8,10 @@ import (
 type Post struct {
 
 	// Идентификатор данного сообщения.
-	Id float32 `json:"id,omitempty"`
+	Id int `json:"id,omitempty"`
 
 	// Идентификатор родительского сообщения (0 - корневое сообщение обсуждения).
-	Parent float32 `json:"parent,omitempty"`
+	Parent int `json:"parent,omitempty"`
 
 	// Автор, написавший данное сообщение.
 	Author string `json:"author"`
@@ -26,7 +26,7 @@ type Post struct {
 	Forum string `json:"forum,omitempty"`
 
 	// Идентификатор ветви (id) обсуждения данного сообещния.
-	Thread float32 `json:"thread,omitempty"`
+	Thread int `json:"thread,omitempty"`
 
 	// Дата создания сообщения на форуме.
 	Created time.Time `json:"created,omitempty"`
